requirements.txt
playwright>=1.44.0
openai>=1.30.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pillow>=10.0.0
imagehash>=4.3.1
httpx>=0.27.0
